Conflicts:
	src/main/java/com/lucazz82/task/models/UserModel.java

+++ src/main/java/com/lucazz82/task/repositories/TaskRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.lucazz82.task.models.TaskModel;
 
-public interface TaskRepository extends CrudRepository<TaskModel, Integer>{
+public interface TaskRepository extends CrudRepository<TaskModel, Long>{
 	public ArrayList<TaskModel> findAll();
 	public Optional<TaskModel> findById(Long id);
 }
